requirements.txt

matplotlib==3.10.1
numpy==2.2.0
usgs-libcomcat==2.1.6
DateTime==5.5
statistics==3.4
scipy==1.15.2
plotly==6.0.1
xarray==v2025.03.0
dash==3.0.1

